backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dashscope==1.23.5
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
Flask==3.1.1
flask-cors==6.0.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.0
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
multidict==6.5.0
mysql-connector-python==9.3.0
numpy==2.2.6
openai==1.90.0
pandas==2.3.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
qwen-agent==0.0.27
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.25.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
yarl==1.20.1
